requirements.txt
colorama==0.4.3
Django==2.2.9
django-service-objects==0.5.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
flake8-polyfill==1.0.2
future==0.18.2
isort==4.3.21
mando==0.6.4
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
radon==4.0.0
six==1.13.0
sqlparse==0.3.0
